Converted separate binaries into instances of AltctrlInterface trait
